Clean up a few files
